Enhance responsive design with media queries
Added media queries to `hobbies.component.css` to adjust the width of the `.HobbiesContainer` and the grid layout of its child elements for screen widths of 1400px, 1000px, and 600px.

Modified `introduction.component.css` by changing the padding of the `.introductionSection` from `10px 0` to `100px 0` and removing the fixed height of `350px`. Also removed the height auto adjustment for `.introductionSection` in the media query for screens with a max-width of 600px.

Updated `projects.component.css` by removing the fixed height of `350px` from `.ProjectsSection` and adding padding-bottom of `20px`. Added media queries to adjust the width of the `.ProjectsContainer` for screens with max-widths of 900px and 600px.

Added media queries to `work-experience.component.css` to adjust the width of the `.WorkExperience` and `.Bijbaantjes` containers for screens with max-widths of 1150px, 800px, and 500px.