Add GetContacts method and remove debug statement
Removed the `Console.WriteLine("dddddddddddddddd");` statement from the `UploadContact` method in the `ContactController` class. Added a new asynchronous `GetContacts` method to the `ContactController` class, which retrieves a list of contacts using the `_contactService.GetContacts()` method and returns the result with an `Ok` response. The `GetContacts` method is decorated with `[HttpGet("contacts")]` for routing and `[ContactListFilter]` for filtering or authorization.